# make `test_dumpRecoGeometry` run faster