@cmsbuild 2024-09-20
A new Pull Request was created by  @mmusich for master.

It involves the following packages:

- Fireworks/Geometry (**visualization**)


@Dr15Jones, @alja, @cmsbuild, @makortel can you please review it and eventually sign? Thanks.
@alja this is something you requested to watch as well.
@antoniovilela, @mandrenguyen, @rappoccio, @sextonkennedy you are the release manager for this.

cms-bot commands are listed <a href="http://cms-sw.github.io/cms-bot-cmssw-cmds.html">here</a>